all non-api pages moved to app router
