+++ app/admin/add_lesson/page.js
@@ -1,0 +1,8 @@
+import AddDesc from "./add-lesson-page";
+import { fetchDBStructure } from "../../../middleware";
+
+export default async function Page() {
+    const db = await fetchDBStructure({})
+    return <AddDesc db={db} />
+}
+
