+++ app/admin/add_q/AddQ.module.css
@@ -1,0 +1,5 @@
+.input {
+    padding: 5px 5px;
+    margin: 0%;
+    text-align: center;
+}
