+++ app/profile/page.js
@@ -1,0 +1,9 @@
+import Profile from './profile';
+import { getServerSession } from "next-auth/next";
+import { authOptions } from "../../app/api/auth/[...nextauth]/route";
+
+export default async function Page({ searchParams }) {
+    const { user } = await getServerSession(authOptions)
+
+    return <Profile user={user} ID={searchParams.ID} />
+}
