+++ app/profile/vocab-list/Vocab.module.css
@@ -1,0 +1,6 @@
+.image {
+    display: inline-flex;
+    justify-content: space-between;
+    gap: 15px;
+    align-items: center;
+}
