+++ app/study/[sphere]/page.js
@@ -1,9 +1,10 @@
 
 import React from "react";
-import { MainContainer, Grid, GridCard } from '../../../components/meta'
+import { Grid, GridCard } from '../../../components/meta'
 import { getServerSession } from "next-auth/next";
-import { fetchDBStructure } from "../../../middleware";
 import { authOptions } from "../../api/auth/[...nextauth]/route";
+import { fetchDBStructure } from "../../../middleware";
+import { Intro, Title } from '../../../components/layout'
 
 export default async function SpherePage({ params }) {
     const db = await fetchDBStructure({ sphere: params.sphere })
@@ -19,29 +20,26 @@ export default async function SpherePage({ params }) {
 
     return (
         <>
-            <MainContainer
-                navType='other'
-                titleText={db[0].sphere}
-                introText={db[0].description}
-            >
-                <Grid>
-                    {
-                        db[0].courses.map(e => {
-                            return (
-                                <GridCard
-                                    key={e._id}
-                                    isDisabled={!e.lessons.length}
-                                    link={`/study/${db[0].slug}/${e.slug}?ID=${ID}`}
-                                    title={e.course}
-                                    description={e.description}
-                                />
 
-                            )
-                        })
-                    }
+            <Title>{db[0].sphere}</Title>
+            <Intro>{db[0].description}</Intro>
+            <Grid>
+                {
+                    db[0].courses.map(e => {
+                        return (
+                            <GridCard
+                                key={e._id}
+                                isDisabled={!e.lessons.length}
+                                link={`/study/${db[0].slug}/${e.slug}?ID=${ID}`}
+                                title={e.course}
+                                description={e.description}
+                            />
 
-                </Grid>
-            </MainContainer>
+                        )
+                    })
+                }
+
+            </Grid>
         </>
     )
 }
