+++ components/layout/Intro/Intro.module.css
@@ -1,0 +1,6 @@
+.intro {
+    margin-bottom: 25px;
+    font-size: 1.25rem;
+    max-width: 850px;
+    text-align: center;
+}
