+++ components/layout/Intro/index.js
@@ -1,0 +1,2 @@
+export * from './Intro';
+export { default } from './Intro';
