+++ components/layout/MainContainer/MainContainer.module.css
@@ -1,0 +1,13 @@
+.container {
+    padding: 0 2rem;
+}
+
+.main {
+    min-height: 85vh;
+    padding: 4rem 0;
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    align-items: center;
+}
