+++ components/layout/Title/Title.module.css
@@ -1,0 +1,13 @@
+.bigTitle {
+    text-align: center;
+    margin: 30px auto;
+    line-height: 1.15;
+    font-size: 4rem;
+}
+
+.smallTitle {
+    text-align: center;
+    margin: 0px auto;
+    line-height: 1.15;
+    font-size: 2.5rem;
+}
