+++ components/layout/index.js
@@ -1,8 +1,13 @@
 import Footer from "./Footer";
 import NavBar from "./NavBar";
+import MainContainer from "./MainContainer";
+import Title from './Title'
+import Intro from "./Intro";
 
 export {
     Footer,
     NavBar,
-
+    MainContainer,
+    Title,
+    Intro
 }
