--- components/meta/Loading/index.js
@@ -1,2 +1,0 @@
-export * from './Loading';
-export { default } from './Loading';
