+++ components/meta/index.js
@@ -1,15 +1,11 @@
 
 import Grid from "./Grid"
 import GridCard from "./GridCard";
-import MainContainer from "./MainContainer";
-import Loading from './Loading';
 import AuthBlock from './AuthBlock'
 
 export {
     Grid,
     GridCard,
-    MainContainer,
-    Loading,
     AuthBlock,
 
 }
