--- pages/404.js
@@ -1,21 +1,0 @@
-import React from "react";
-import { MainContainer } from "../components/meta";
-import { BlueButton } from "../components/atomic";
-
-const FourOhFour = () => {
-    return (
-        <MainContainer
-            titleText='404'
-            introText="This page doesn't exist... yet"
-        >
-            <p>Do you want it to?</p>
-            <BlueButton
-                id='404'
-                onClick={() => alert('duly noted')}
-            >Yes</BlueButton>
-        </MainContainer>
-    )
-}
-
-export default FourOhFour
-
